add fetchStyles utility
